requirements.txt
anyio==3.6.1
click==8.1.3
coloredlogs==15.0.1
fastapi==0.85.0
flatbuffers==2.0.7
h11==0.13.0
httptools==0.5.0
humanfriendly==10.0
idna==3.4
mpmath==1.2.1
numpy==1.23.3
onnx==1.12.0
onnxruntime==1.12.1
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
protobuf==3.20.1
pydantic==1.10.2
pyparsing==3.0.9
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
sympy==1.11.1
typing_extensions==4.3.0
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
